Progress, but i realized its not want I want to do
Made a AJAX function that fetches marker information from the database
which is fine, the problem is, we also need to be able to display this
information on the page (like  the display page for assignment 2) so the
display.php page will need to access the database anyway to create a
dynamic page with marker information on it (like in the sidebar or
whatever) so it will be repeat functionality. chances are the best way
to do this will be do move the PHP from getPins.php to the display page,
and use it to generate the DOM and then pass the JSON object to the
javascript so it can do the google maps stuff.